refactor: main.goをコマンドごとのファイルに分割
- main.goを config.go / gist.go / cmd_*.go に分割
- *viper.Viper を各関数に注入してテスタビリティを向上
- ユニットテストを追加 (testify)
- syncToGist のバックグラウンドプロセス起動失敗を warn 出力
- listProfiles のループ変数 shadowing を解消
- viperInstance() を cmd_root.go に移動
- CI ワークフロー (ci.yml) を追加
